unify fake mailers
